Trying again

+++ cpp/src_prims/selection/knn.h
@@ -140,7 +140,7 @@ void brute_force_knn(float **input, int *sizes, int n_params, IntType D,
     std::cout << "Using translations: [" << std::endl;
     for (int i = 0; i < translations->size(); i++) {
       int64_t s = translations[i];
-      std::cout << s << ", ";
+      std::cout << (int)s << ", ";
     }
 
     std::cout << "]" << std::endl;
